feat!: Move `Configuration` and `VerticleConfig` data types to application module
BREAKING CHANGE: Remove `Configuration` and `VerticleConfig` from services module and add them to the application module.
