+++ modules/telestion-application/src/main/java/de/wuespace/telestion/application/Telestion.java
@@ -8,7 +8,6 @@ import io.vertx.core.Vertx;
 import java.util.Collections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import de.wuespace.telestion.services.config.Configuration;
 
 /**
  * This is going to be the Telestion application. It launches the verticles which are specified in the configuration.
