+++ modules/telestion-application/src/main/java/de/wuespace/telestion/application/VerticleConfig.java
@@ -1,4 +1,4 @@
-package de.wuespace.telestion.services.config;
+package de.wuespace.telestion.application;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
